add a .gitignore and requirements
so a '.gitignore' file is pretty simple but it helps you leave out certain files that you don't always want to push up to github. for that .idea folder, that is just something PyCHarm is creating so it can store your settings correctly, so we can ignore that since when I download your code my .idea file is going to be used. So when it comes to that venv folder, there is no reason to push our entire Python environment everytime we submit changes, so .gitignore that.  What we do instead, is  create a requirements.txt file. When you develop this project out more and possibly use selenium, or some other package that doesn't come with basic python you state it there. That way someone can just look through the requirements.txt file and see that you are using selenium instead of sifting through the entire venv folder. I put selenium in there as an example, you can delete it for now if you want.