add a main file
So even more so than the last commit, we are going to create a full file called main that will 1. make it explicit on where the program starts and 2. give a central location to import  our files. I notices in the Macros file you had a weird import happening since you were just trying to download specific functions from the other file.