update to gitignore and use function parameters
like in sql, the '*'  is a wildcard. So, instead of having to delete the .pyc files that keep poppiing up (if you want me to explain what those are I can) we are just going to use the wildcard to ignore any file that has the .pyc ending. Next, I used function parameters for the bmr functions to pass in values.